requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-telegram-bot==20.7
httpx==0.25.2
python-dotenv==1.0.0
